add exception

+++ src/main/java/aibles/userprofilemanager_1/service/serviceImpl/PostServiceImpl.java
@@ -3,6 +3,7 @@ package aibles.userprofilemanager_1.service.serviceImpl;
 import aibles.userprofilemanager_1.dto.PostRequest;
 import aibles.userprofilemanager_1.dto.PostResponse;
 import aibles.userprofilemanager_1.entity.PostEntity;
+import aibles.userprofilemanager_1.exception.NotFoundException;
 import aibles.userprofilemanager_1.repository.PostRepository;
 import aibles.userprofilemanager_1.service.mapping.PostMapping;
 import aibles.userprofilemanager_1.service.service.PostService;
@@ -11,72 +12,56 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
 public class PostServiceImpl implements PostService {
     private final PostRepository postRepository;
+
     @Autowired
-    public PostServiceImpl(PostRepository postRepository)
-    {
-        this.postRepository=postRepository;
+    public PostServiceImpl(PostRepository postRepository) {
+        this.postRepository = postRepository;
     }
 
     @Override
     @Transactional
-    public PostResponse create(PostRequest request)
-    {
-        PostEntity entity= PostMapping.convertDtoEntity(request);
-        PostEntity saveEntity=postRepository.save(entity);
-        return PostMapping.convertEntityToPostResponse(saveEntity);
+    public PostResponse create(PostRequest request) {
+        PostEntity entity = PostMapping.convertDtoEntity(request);
+        PostEntity savedEntity = postRepository.save(entity);
+        return PostMapping.convertEntityToPostResponse(savedEntity);
     }
 
     @Override
-    @org.springframework.transaction.annotation.Transactional(readOnly = true)
+    @Transactional
     public List<PostResponse> getAllPosts() {
         List<PostEntity> entities = postRepository.findAll();
         return entities.stream().map(PostMapping::convertEntityToPostResponse).collect(Collectors.toList());
     }
 
     @Override
-    //@Transactional(readOnly = true)
-    public PostResponse getPostById(Long id)
-    {
-        Optional<PostEntity> entity=postRepository.findById(id);
-        if(entity.isPresent())
-        {
-            return PostMapping.convertEntityToPostResponse(entity.get());
-        }
-        else {
-            throw new RuntimeException("Post not found with id: "+id);
-        }
+    public PostResponse getPostById(Long id) {
+        return postRepository.findById(id)
+                .map(PostMapping::convertEntityToPostResponse)
+                .orElseThrow(() -> new NotFoundException("Post not found with id: " + id));
     }
 
     @Override
     @Transactional
-    public PostResponse updatePosts(Long id,PostRequest request)
-    {
-        Optional<PostEntity> existingEntity = postRepository.findById(id);
-        if(existingEntity.isPresent())
-        {
-            PostEntity updatedEntity = existingEntity.get();
-            updatedEntity.setTitle(request.getTitle());
-            updatedEntity.setContent(request.getContent());
-            PostEntity savedEntity = postRepository.save(updatedEntity);
-            return PostMapping.convertEntityToPostResponse(savedEntity);
-        }
-        else
-        {
-            throw new RuntimeException("Post not found with id: "+id);
-        }
+    public PostResponse updatePosts(Long id, PostRequest request) {
+        PostEntity existingEntity = postRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("Post not found with id: " + id));
+        existingEntity.setTitle(request.getTitle());
+        existingEntity.setContent(request.getContent());
+        PostEntity savedEntity = postRepository.save(existingEntity);
+        return PostMapping.convertEntityToPostResponse(savedEntity);
     }
 
     @Override
     @Transactional
-    public void deletePost(Long id)
-    {
+    public void deletePost(Long id) {
+        if (!postRepository.existsById(id)) {
+            throw new NotFoundException("Post not found with id: " + id);
+        }
         postRepository.deleteById(id);
     }
-
 }
